chore: move client-go features configuration to util/settings/settings.go
