+++ util/settings/cluster_informer.go
@@ -13,7 +13,6 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	clientgofeatures "k8s.io/client-go/features"
 	informersv1 "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
@@ -277,32 +276,4 @@ func secretToCluster(s *corev1.Secret) (*appv1.Cluster, error) {
 	cluster.Namespace = s.Namespace
 
 	return &cluster, nil
-}
-
-func init() {
-	ConfigureGoClientFeatures()
-}
-
-type myFeatureGates struct {
-	parent clientgofeatures.Gates
-}
-
-func (m myFeatureGates) Enabled(f clientgofeatures.Feature) bool {
-	if f == clientgofeatures.WatchListClient ||
-		f == clientgofeatures.InOrderInformers {
-		return false
-	}
-	return m.parent.Enabled(f)
-}
-
-// FIXME: remove when we have proper WatchListClient and InOrderInformers support
-func ConfigureGoClientFeatures() {
-	gates := clientgofeatures.FeatureGates()
-	isWatchListEnabled := gates.Enabled(clientgofeatures.WatchListClient)
-	isInOrderInformersEnabled := gates.Enabled(clientgofeatures.InOrderInformers)
-
-	if isWatchListEnabled || isInOrderInformersEnabled {
-		wrapper := myFeatureGates{parent: gates}
-		clientgofeatures.ReplaceFeatureGates(wrapper)
-	}
 }
